Fix MinGW and MSVC build errors and warnings
- Add MinGW SDL2 support via find_package and proper library linking
- Fix crashhandler Windows include order (windows.h before dbghelp.h)
- Wrap mmap/munmap calls with #ifndef _WIN32 for Windows compatibility
- Fix undefined behavior in union initialization (float.h, Color.h)
- Add matching operator new to RefCounted to fix mismatched new/delete warning
- Fix EASTL tuple.h tautological compare warning
- Add [[maybe_unused]] to g_shortcuts_list to suppress unused variable warning
- Fix cfgfile_searchconfig bounds check for empty input strings
- Remove hardcoded NDEBUG from ADFlib source files (CMake handles this)
- Guard GCC-specific pragmas with compiler checks for cross-platform compatibility
- Fix calc.cpp switch on high-bit TCHAR operators with proper cast
- Add ws2_32 library linking for network functions

+++ libs/amDebugger/src/amDebugger/window/mmu_wnd.cpp
@@ -92,7 +92,7 @@ void MmuWnd::drawContentImp() {
         ImGui::TableHeadersRow();
 
         ImGuiListClipper clipper;
-        clipper.Begin(m_filteredPages.size());
+        clipper.Begin(static_cast<int>(m_filteredPages.size()));
         while (clipper.Step()) {
             for (int row_n = clipper.DisplayStart; row_n < clipper.DisplayEnd; row_n++) {
                 const auto& page = m_filteredPages[row_n];
